+++ src/backend/common/Threads.cpp
@@ -136,8 +136,6 @@ xmrig::String xmrig::Threads<T>::profileName(const Algorithm &algorithm, bool st
         }
     }
 
-	if (name == "defyx" && has("rx")) return "rx";
-
     if (has(kAsterisk)) {
         return kAsterisk;
     }