+++ src/crypto/common/Algorithm.cpp
@@ -112,8 +112,11 @@ static AlgoName const algorithm_names[] = {
     { "RandomXL",                  nullptr,            Algorithm::RX_LOKI         },
     { "randomx/arq",               "rx/arq",           Algorithm::RX_ARQ          },
     { "RandomARQ",                 nullptr,            Algorithm::RX_ARQ          },
+    { "randomx/sfx",               "rx/sfx",           Algorithm::RX_SFX          },
+    { "RandomSFX",                 nullptr,            Algorithm::RX_SFX          },
+    { "randomx/v",                 "rx/v",             Algorithm::RX_V            },
+    { "RandomV",                   nullptr,            Algorithm::RX_V            },
     { "DefyX",                     "defyx",            Algorithm::DEFYX           },
-	
 #   endif
 #   ifdef XMRIG_ALGO_ARGON2
     { "argon2/chukwa",             nullptr,            Algorithm::AR2_CHUKWA      },
@@ -140,17 +143,17 @@ size_t xmrig::Algorithm::l2() const
     switch (m_id) {
     case RX_0:
     case RX_LOKI:
+    case RX_SFX:
+    case RX_V:
         return 0x40000;
 
     case RX_WOW:
+    case DEFYX:
         return 0x20000;
 
     case RX_ARQ:
         return 0x10000;
 
-    case DEFYX:
-        return 0x20000;
-		
     default:
         break;
     }
@@ -178,17 +181,17 @@ size_t xmrig::Algorithm::l3() const
         switch (m_id) {
         case RX_0:
         case RX_LOKI:
+        case RX_SFX:
+        case RX_V:
             return oneMiB * 2;
 
         case RX_WOW:
             return oneMiB;
 
         case RX_ARQ:
-            return oneMiB / 4;
-
         case DEFYX:
             return oneMiB / 4;
-			
+
         default:
             break;
         }
@@ -280,6 +283,8 @@ xmrig::Algorithm::Family xmrig::Algorithm::family(Id id)
     case RX_WOW:
     case RX_LOKI:
     case RX_ARQ:
+    case RX_SFX:
+    case RX_V:
     case DEFYX:	
         return RANDOM_X;
 #   endif