+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+# v5.0.1
+
+- Rebase from XMRig v5.4.0 [(changelog)](doc/CHANGELOG.md)
+
+
 # v5.0.0
 
-- Rebase from XMRig v5.3.0 [(changelog)](doc/CHANGELOG.md)
+- Rebase from XMRig v5.3.0
 - Some extra DefyX optimizations
 - Disabled OpenCL/Cuda
 - Updated Readme and changelog